remove data_pipeline & replace files to util
