--- zae_engine/data_pipeline/__init__.py
@@ -1,1 +1,0 @@
-from ._fileio import example_ecg, example_mri
